moving check_gsim_list to sm_utils
